+++ Nemesis-Mod/Nemesis-Mod.ino
@@ -1,8 +1,8 @@
 #include <Arduino.h>
-#include "src/hardware/Mainboard.h"
 #include "src/App.h"
 #include "src/Button.h"
 #include "src/Log.h"
+#include "src/Mainboard.h"
 
 const uint16_t CLEAR_HOLD_IN_MSECS = 30000;  // 30 seconds
 const uint16_t RESET_HOLD_IN_MSECS = 5000;   // 5 seconds
